migrate get-megamenu
